+++ lib/get_apps.dart
@@ -2,21 +2,33 @@ import 'get_apps_platform_interface.dart';
 import 'models.dart';
 
 class GetApps {
-
-  Future<List<AppInfo>> getApps({bool includeSystemApps = false}) {
-    return GetAppsPlatform.instance.getApps(includeSystemApps: includeSystemApps);
+  Future<AppInfo> getAppInfo(
+    String packageName, {
+    bool shouldInitialize = false,
+  }) {
+    return GetAppsPlatform.instance.getAppInfo(
+      packageName,
+      shouldInitialize: shouldInitialize,
+    );
   }
 
-  Future<void> openApp(String packageName) async{
+  Future<List<AppInfo>> getApps({bool includeSystemApps = false}) {
+    return GetAppsPlatform.instance.getApps(
+      includeSystemApps: includeSystemApps,
+    );
+  }
+
+  Future<void> openApp(String packageName) async {
     return GetAppsPlatform.instance.openApp(packageName);
   }
 
-  Future<void> deleteApp(String packageName) async{
+  Future<void> deleteApp(String packageName) async {
     return GetAppsPlatform.instance.deleteApp(packageName);
   }
 
   Stream<ActionNotification> appActionReceiver() async* {
-    await for (final actionNotification in GetAppsPlatform.instance.appActionReceiver()){
+    await for (final actionNotification
+        in GetAppsPlatform.instance.appActionReceiver()) {
       yield actionNotification;
     }
   }